fix: logbook and external certificates for mobile
