+++ apps/web/src/app/(dashboard)/(account)/profiel/[handle]/_components/external-certificate/certificate-templates.ts
